Reorganize Apify-related files into posts subdirectory and update imports
